+++ resources/views/admin/roles/index.blade.php
@@ -13,20 +13,26 @@
     <!-- Striped Rows -->
     @include('admin.partials.content-header', ['pages' => 'Account Manage', 'name' => 'Roles'])
     <ul class="nav nav-pills flex-column flex-md-row mb-3">
-        <li class="nav-item">
-            <a class="nav-link active" href="javascript:void(0);"><i class="bx bx-shield me-1"></i> List Roles</a>
-        </li>
-        <li class="nav-item">
-            <a class="nav-link" href="{{route('roles.create')}}">
-                <i style="font-size: 24px;" class="bx bx-plus-circle me-1"></i>Create Roles
-            </a>
-        </li>
-        <li class="nav-item">
-            <button id="deleteAllSelectedRecord" class="btn btn-outline-danger"
-                    data-url="{{ route('roles.deleteSelected')}}">
-                <i class="bx bx-trash me-1"></i>Delete Roles
-            </button>
-        </li>
+        @include('admin.partials.content-body', [
+           'permissions_list' => 'list-roles',
+           'active_list' => 'active',
+           'models' => 'Roles',
+           'url_list' => route('roles.index'),
+           'class_list' => 'bx bx-shield me-1',
+           'permission_create' => 'create-role',
+           'active_create' => '',
+           'model' => 'Role',
+           'url_create' => route('roles.create'),
+           'class_create' => 'bx bx-plus-circle me-1'
+       ])
+        @if(auth()->user()->checkPermissionAccess('delete-role') == true)
+            <li class="nav-item">
+                <button id="deleteAllSelectedRecord" class="btn btn-outline-danger"
+                        data-url="{{ route('roles.deleteSelected')}}">
+                    <i class="bx bx-trash me-1"></i>Delete Roles
+                </button>
+            </li>
+        @endif
     </ul>
     <div class="card">
         <h5 class="card-header">List Roles</h5>
